+++ ReleaseNotes.md
@@ -4,6 +4,7 @@
 
 ## Bugfix
 
+* TGR-430: Requests without Deep-Paths are now forwarded from remote Tiger-Proxies
 * TGR-424: Disabled Rbel-Parsing will no longer impact traffic forwarding
 * TGR-422: Find request to path "http://server" now also matches "/"
 * TGR-427: JSON-Checker now correctly reports mismatched types and ill-formatted JSONs